+++ src/app/modules/project/project.module.ts
@@ -14,7 +14,6 @@
  *   along with this program, in the LICENSE.md file in the root project directory.
  *   If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
  */
-
 import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 import { DetailsComponent } from "./details/details.component";
@@ -23,9 +22,11 @@ import { ProjectRoutingModule } from "./project-routing.module";
 import { AccordionModule } from "ngx-bootstrap/accordion";
 import { EditComponent } from "./edit/edit.component";
 import { ReactiveFormsModule, FormsModule } from "@angular/forms";
+import { EmbedButtonComponent } from './embed-button/embed-button.component';
+import { EmbedComponent } from './embed/embed.component';
 
 @NgModule({
-  declarations: [OverviewComponent, DetailsComponent, EditComponent],
+  declarations: [OverviewComponent, DetailsComponent, EditComponent, EmbedButtonComponent, EmbedComponent],
   imports: [CommonModule, FormsModule, ReactiveFormsModule, ProjectRoutingModule, AccordionModule.forRoot()],
 })
 export class ProjectModule {}